requirements.txt
Flask==2.3.3
openai==1.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27
requests==2.28.2
Werkzeug==2.3.7